reorganize the project's file, directory layout
